requirements.txt
packaging==26.0
pycryptodome==3.23.0
setuptools==82.0.1
wheel==0.46.3
